+++ Src/LineBvh.fs
@@ -101,7 +101,7 @@ type LineBvh private (bvh: Bvh<Line3D>) =
     /// <summary>Finds the indices of all lines whose bounding box is closer to the given
     /// axis aligned bounding box than the given tolerance.</summary>
     /// <param name="box">The axis aligned bounding box to search in.</param>
-    /// <param name="tolerance">The tolerance distance around the box. Optional, 0.0 by default.</param>
+    /// <param name="tolerance">The tolerance distance around the box. Must not be negative. Optional, 0.0 by default.</param>
     /// <returns>A ResizeArray of the indices of the found lines in the input array.</returns>
     member _.LinesInBox (box: BBox, [<OPT;DEF(0.0)>] tolerance: float) : ResizeArray<int> =
         bvh.ItemsInBox (box, tolerance)
@@ -109,7 +109,7 @@ type LineBvh private (bvh: Bvh<Line3D>) =
     /// <summary>Finds the indices of all lines whose bounding box is closer to the given
     /// 3D point than the given tolerance.</summary>
     /// <param name="pt">The 3D point to search around.</param>
-    /// <param name="tolerance">The tolerance distance around the point. Optional, 0.0 by default.</param>
+    /// <param name="tolerance">The tolerance distance around the point. Must not be negative. Optional, 0.0 by default.</param>
     /// <returns>A ResizeArray of the indices of the found lines in the input array.</returns>
     member _.LinesNearPoint (pt: Pnt, [<OPT;DEF(0.0)>] tolerance: float) : ResizeArray<int> =
         bvh.ItemsNearPoint (pt, tolerance)